+++ spinn_front_end_common/utilities/utility_objs/extra_monitor_scp_processes/__init__.py
@@ -16,14 +16,19 @@
 from .read_status_process import ReadStatusProcess
 from .reset_counters_process import ResetCountersProcess
 from .set_packet_types_process import SetPacketTypesProcess
-from .set_router_wait1_timeout_process import SetRouterTimeoutProcess
+from .set_router_timeout_process import SetRouterTimeoutProcess
 from .set_router_wait2_timeout_process import SetRouterEmergencyTimeoutProcess
 from .clear_queue_process import ClearQueueProcess
 from .load_application_mc_routes_process import LoadApplicationMCRoutesProcess
 from .load_system_mc_routes_process import LoadSystemMCRoutesProcess
 
 __all__ = [
-    "ReadStatusProcess", "ResetCountersProcess", "SetPacketTypesProcess",
-    "SetRouterEmergencyTimeoutProcess", "SetRouterTimeoutProcess",
-    "ClearQueueProcess", "LoadApplicationMCRoutesProcess",
-    "LoadSystemMCRoutesProcess"]
+    "ClearQueueProcess",
+    "LoadApplicationMCRoutesProcess",
+    "LoadSystemMCRoutesProcess",
+    "ReadStatusProcess",
+    "ResetCountersProcess",
+    "SetPacketTypesProcess",
+    "SetRouterEmergencyTimeoutProcess",
+    "SetRouterTimeoutProcess",
+    ]
